Add input slider legend, move input components
