+++ src/components/Form/InputMoves.js
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react'
 import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
-import { ReactComponent as ArrowRightIcon } from '../assets/img/arrow_right.svg'
-import { ReactComponent as ArrowDownIcon } from '../assets/img/arrow_down.svg'
+import { ReactComponent as ArrowRightIcon } from '../../assets/img/arrow_right.svg'
+import { ReactComponent as ArrowDownIcon } from '../../assets/img/arrow_down.svg'
 
 InputMoves.propTypes = {
   levelName: PropTypes.string.isRequired,
